UI-7461 - Inline the individual legacy migration in migrateCalculatedMeasures.
